rename openapi package -> beacon
