+++ api/client/beacon/doc.go
@@ -3,4 +3,4 @@ Package openapi provides a client for interacting with the standard Eth Beacon N
 Interactive swagger documentation for the API is available here: https://ethereum.github.io/beacon-APIs/
 
 */
-package openapi
+package beacon
