+++ api/client/beacon/errors.go
@@ -1,4 +1,4 @@
-package openapi
+package beacon
 
 import "github.com/pkg/errors"
