+++ beacon-chain/sync/checkpoint/BUILD.bazel
@@ -9,7 +9,7 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/beacon-chain/sync/checkpoint",
     visibility = ["//visibility:public"],
     deps = [
-        "//api/client/openapi:go_default_library",
+        "//api/client/beacon:go_default_library",
         "//beacon-chain/db:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
