+++ cmd/prysmctl/checkpoint/BUILD.bazel
@@ -10,7 +10,7 @@ go_library(
     importpath = "github.com/prysmaticlabs/prysm/cmd/prysmctl/checkpoint",
     visibility = ["//visibility:public"],
     deps = [
-        "//api/client/openapi:go_default_library",
+        "//api/client/beacon:go_default_library",
         "@com_github_sirupsen_logrus//:go_default_library",
         "@com_github_urfave_cli_v2//:go_default_library",
     ],
